Structure some Wagtail code
