+++ src/main/blocks/__init__.py
@@ -1,1 +1,2 @@
 from .text_and_media import *  # NOQA
+from .scenario import *  # NOQA
