+++ src/main/pages/storyline.py
@@ -5,54 +5,22 @@ from wagtail_headless_preview.models import HeadlessPreviewMixin
 from wagtail.admin.panels import FieldPanel
 from wagtail.fields import StreamField
 
-from api.models import Slider
 
 from .base import BasePage
-from ..blocks import TextAndMediaBlock
-
-scenarios = [
-    ("scenario-1", "Scenario 1"),
-    ("scenario-2", "Scenario 2"),
-    ("scenario-3", "Scenario 3"),
-]
-
-
-def get_sliders():
-    return [(slider.pk, slider.name) for slider in Slider.objects.all()]
-
-
-class SliderBlock(blocks.StructBlock):
-    slider = blocks.ChoiceBlock(choices=get_sliders)
-    visible = blocks.BooleanBlock(required=False)
+from ..blocks import TextAndMediaBlock, ScenarioBlock
 
 
 class StorylinePage(HeadlessPreviewMixin, BasePage):
     storyline = StreamField(
         [
             ("intro", TextAndMediaBlock()),
-            (
-                "section",
-                blocks.StructBlock(
-                    [
-                        ("scenario", blocks.ChoiceBlock(choices=scenarios)),
-                        (
-                            "content",
-                            blocks.StreamBlock(
-                                [
-                                    ("text", blocks.RichTextBlock()),
-                                    ("slider", SliderBlock()),
-                                    # ("radiobuttons", RadioButtonBlock()),
-                                ]
-                            ),
-                        ),
-                    ]
-                ),
-            ),
+            ("section", ScenarioBlock()),
         ],
         block_counts={"intro": {"min_num": 1, "max_num": 1}},
         use_json_field=True,
     )
 
+    serializer_class = "main.pages.StorylinePageSerializer"
     content_panels = BasePage.content_panels + [
         FieldPanel("storyline"),
     ]
